requirements.txt
pycryptodome==3.9.9
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
